+++ public/app/oo/launch-python-server.sh
@@ -1,1 +1,2 @@
-python -m SimpleHTTPServer 9001
+#python -m SimpleHTTPServer 9001
+python3 -m http.server 9001